Adding a single flow for doc upload and add to corpus.
